requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytest>=8.0.0
pytest-cov>=5.0.0
python-dotenv>=1.0.0
tabulate>=0.9.0
